Update README.md
Write a function that draws a line following coordinates. For this lab, there is no skeleton file provided, so you have to write your main function that interacts with files (descriptions to follow).

You are given an input file which contains a set of (x, y) coordinates. These coordinates are written, so that a line segment is connected from the coordinate in the preceding line to the coordinate. For example, let’s assume that input.txt contains the following coordinates:

    cat input.txt

0,0

0,1

1,1

1,0

0,0